+++ ptsa/extensions/morlet/morlet.cpp
@@ -27,27 +27,52 @@ size_t nextpow2(size_t v) {
 }
 
 
-size_t MorletWaveFFT::init(size_t width, double freq, size_t win_size, double sample_freq) {
+size_t MorletWaveFFT::init(size_t width, double freq, size_t win_size, double sample_freq, bool complete) {
     double dt = 1.0 / sample_freq;
     double sf = freq / width; //sigma_f;  width of Gaussian in the frequency domain
     double st = 1.0 / (2.0 * M_PI * sf); //sigma_t; width of Gaussian in the time domain.
-    double a = 1 / sqrt(st * sqrt(M_PI));
+    double a_c = 1 / sqrt(st * sqrt(M_PI));
+    double a_s = a_c;
     double omega = 2.0 * M_PI * freq;
 
-    nt = size_t(7.0 * st / dt) + 1;
+    double sample_factor = 10.0;
+    nt = size_t(sample_factor * st / dt) + 1;
+
+    double t = -(sample_factor/2.0) * st;
+    double scale = 2.0 * st * st;
+    double complete_offset = 0;
+    double freq_scale = 1;
+
+    if (complete) {
+      complete_offset = exp(-(double)width*width/2.0);
+      freq_scale = (2/M_PI)*(acos(exp(-0.5*width*width)));
+      nt = size_t((nt-1)/freq_scale + 1.5);
+      t = t/freq_scale;
+      scale /= freq_scale*freq_scale;
+      double inv_sq_a_c = (1.0 / freq_scale) * ( width/(4.0*freq*sqrt(M_PI)) +
+          3.0*width*exp(-(double)width*width) / (4.0*freq*sqrt(M_PI)) -
+          (double)width*exp(-3*(double)width*width/4.0) / (freq*sqrt(M_PI)) );
+      double acos_term = acos(exp(-(double)width*width/2.0));
+      double inv_sq_a_s = (width*sqrt(M_PI) / (8*freq *
+          acos(exp(-(double)width*width/2.0)))) *
+          (1-exp(-(double)width*width*M_PI*M_PI/(4*acos_term*acos_term)));
+      a_c = 1.0/sqrt(inv_sq_a_c);
+      a_s = 1.0/sqrt(inv_sq_a_s);
+    }
+
     len0 = win_size + nt - 1;
     len = nextpow2(len0);
     fft = (fftw_complex *) fftw_malloc(len * sizeof(fftw_complex));
 
     fftw_complex *cur_wave = (fftw_complex *) fftw_malloc(len * sizeof(fftw_complex));
 
-    double t = -3.5 * st;
-    double scale = 2.0 * st * st;
     for (size_t i = 0; i < nt; ++i) {
-        double c = a * exp(-t * t / scale);
+        double coef_common = exp(-t * t / scale);
+        double coef_c = a_c * coef_common;
+        double coef_s = a_s * coef_common;
         double omega_t = omega * t;
-        cur_wave[i][0] = c * cos(omega_t);
-        cur_wave[i][1] = c * sin(omega_t);
+        cur_wave[i][0] = coef_c * (cos(freq_scale*omega_t) - complete_offset);
+        cur_wave[i][1] = coef_s * sin(omega_t);
         t += dt;
     }
     for (size_t i = nt; i < len; ++i)
@@ -64,15 +89,15 @@ size_t MorletWaveFFT::init(size_t width, double freq, size_t win_size, double sa
 
 MorletWaveletTransform::MorletWaveletTransform(){}
 
-MorletWaveletTransform::MorletWaveletTransform(size_t width, double *freqs, size_t nf, double sample_freq, size_t signal_len){
-    init_flex(width, freqs, nf, sample_freq,signal_len);
+MorletWaveletTransform::MorletWaveletTransform(size_t width, double *freqs, size_t nf, double sample_freq, size_t signal_len, bool complete){
+    init_flex(width, freqs, nf, sample_freq,signal_len, complete);
 }
 
-MorletWaveletTransform::MorletWaveletTransform(size_t width, double low_freq, double high_freq, size_t nf, double sample_freq, size_t signal_len) {
+MorletWaveletTransform::MorletWaveletTransform(size_t width, double low_freq, double high_freq, size_t nf, double sample_freq, size_t signal_len, bool complete) {
 
     std::vector<double> freqs = logspace(log10(low_freq), log10(high_freq), nf);
 
-    init_flex(width, &freqs[0], nf, sample_freq, signal_len);
+    init_flex(width, &freqs[0], nf, sample_freq, signal_len, complete);
 
 
 }
@@ -98,14 +123,14 @@ MorletWaveletTransform::~MorletWaveletTransform() {
 
 
 void MorletWaveletTransform::init_flex(size_t width, double *freqs, size_t nf, double sample_freq,
-                                       size_t signal_len) {
+                                       size_t signal_len, bool complete) {
     signal_len_ = signal_len;
     n_freqs = nf;
     morlet_wave_ffts = new MorletWaveFFT[nf];
     n_plans = 0;
     size_t last_len = 0;
     for (size_t i = 0; i < nf; ++i) {
-        size_t len = morlet_wave_ffts[i].init(width, freqs[i], signal_len, sample_freq);
+        size_t len = morlet_wave_ffts[i].init(width, freqs[i], signal_len, sample_freq, complete);
         if (len != last_len) {
             last_len = len;
             ++n_plans;
@@ -140,12 +165,12 @@ void MorletWaveletTransform::init_flex(size_t width, double *freqs, size_t nf, d
 }
 
 void MorletWaveletTransform::init(size_t width, double low_freq, double high_freq, size_t nf, double sample_freq,
-                                  size_t signal_len) {
+                                  size_t signal_len, bool complete) {
 
 
     std::vector<double> freqs = logspace(log10(low_freq), log10(high_freq), nf);
 
-    init_flex(width, &freqs[0], nf, sample_freq, signal_len);
+    init_flex(width, &freqs[0], nf, sample_freq, signal_len, complete);
 }